+++ src/plone/app/imagecropping/browser/static/cropping.js
@@ -91,7 +91,6 @@ var ImageCropping = {};
                     selected = $('ul.scales li:first');
                 }
                 var anchor = $(selected).children('a');
-                location.hash = "#" + anchor.attr('id');
                 scales.click(function(e) {
                     if(obj.unsaved_changes()) {
                         return false;